requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
asyncpg
supabase
python-jose[cryptography]
passlib[bcrypt]
cryptography
httpx
tenacity